feat(DTFS2-6967): use kebab-case file names and move type to own file
