+++ dtfs-central-api/api-tests/v1/tfm/deal/tfm-deals-get.api-test.js
@@ -1,4 +1,4 @@
-const { generateTfmAuditDetails, generatePortalAuditDetails } = require('@ukef/dtfs2-common/src/helpers/changeStream/generateAuditDetails');
+const { generateTfmAuditDetails, generatePortalAuditDetails } = require('@ukef/dtfs2-common/src/helpers/change-stream/generate-audit-details');
 const wipeDB = require('../../../wipeDB');
 const app = require('../../../../src/createApp');
 const api = require('../../../api')(app);
