+++ libs/common/src/types/audit-database-record.ts
@@ -1,0 +1,12 @@
+import { ObjectId } from 'mongodb';
+
+/**
+ * The `auditRecord` property on all documents stored in mongodb
+ */
+export type AuditDatabaseRecord = {
+  lastUpdatedAt: Date;
+  lastUpdatedByPortalUserId: ObjectId | null;
+  lastUpdatedByTfmUserId: ObjectId | null;
+  lastUpdatedByIsSystem: boolean | null;
+  noUserLoggedIn: boolean | null;
+};
