+++ trade-finance-manager-api/src/v1/controllers/acbs.controller.js
@@ -1,7 +1,7 @@
 const { ObjectId } = require('mongodb');
 const $ = require('mongo-dot-notation');
 const { DURABLE_FUNCTIONS_LOG } = require('@ukef/dtfs2-common');
-const { generateSystemAuditDatabaseRecord } = require('@ukef/dtfs2-common/src/helpers/changeStream/generateAuditDatabaseRecord')
+const { generateSystemAuditDatabaseRecord } = require('@ukef/dtfs2-common/src/helpers/change-stream/generate-audit-database-record')
 const api = require('../api');
 const db = require('../../drivers/db-client');
 const tfmController = require('./tfm.controller');
