+++ trade-finance-manager-api/src/v1/controllers/deal.controller.js
@@ -1,4 +1,4 @@
-const { generateTfmAuditDetails } = require('@ukef/dtfs2-common/src/helpers/changeStream/generateAuditDetails');
+const { generateTfmAuditDetails } = require('@ukef/dtfs2-common/src/helpers/change-stream/generate-audit-details');
 const mapDeal = require('../mappings/map-deal');
 const api = require('../api');
 const acbsController = require('./acbs.controller');
