+++ trade-finance-manager-api/src/v1/controllers/deal.tasks.test.js
@@ -1,4 +1,4 @@
-const { generatePortalAuditDetails } = require("@ukef/dtfs2-common/src/helpers/changeStream/generateAuditDetails");
+const { generatePortalAuditDetails } = require("@ukef/dtfs2-common/src/helpers/change-stream/generate-audit-details");
 const {
   shouldCreatePartiesTask,
   shouldCreateAgentCheckTask,
